+++ ecd-20200930/src/Models/DescribeCenterPolicyListResponseBody/describePolicyGroups.php
@@ -169,6 +169,11 @@ class describePolicyGroups extends Model
     /**
      * @var string
      */
+    public $internetPrinter;
+
+    /**
+     * @var string
+     */
     public $localDrive;
 
     /**
@@ -364,6 +369,11 @@ class describePolicyGroups extends Model
     /**
      * @var string
      */
+    public $safeMenu;
+
+    /**
+     * @var string
+     */
     public $scope;
 
     /**
@@ -536,6 +546,7 @@ class describePolicyGroups extends Model
         'html5Access' => 'Html5Access',
         'html5FileTransfer' => 'Html5FileTransfer',
         'internetCommunicationProtocol' => 'InternetCommunicationProtocol',
+        'internetPrinter' => 'InternetPrinter',
         'localDrive' => 'LocalDrive',
         'maxReconnectTime' => 'MaxReconnectTime',
         'memoryDownGradeDuration' => 'MemoryDownGradeDuration',
@@ -575,6 +586,7 @@ class describePolicyGroups extends Model
         'resolutionModel' => 'ResolutionModel',
         'resolutionWidth' => 'ResolutionWidth',
         'resourceGroupCount' => 'ResourceGroupCount',
+        'safeMenu' => 'SafeMenu',
         'scope' => 'Scope',
         'scopeValue' => 'ScopeValue',
         'smoothEnhancement' => 'SmoothEnhancement',
@@ -815,6 +827,10 @@ class describePolicyGroups extends Model
             $res['InternetCommunicationProtocol'] = $this->internetCommunicationProtocol;
         }
 
+        if (null !== $this->internetPrinter) {
+            $res['InternetPrinter'] = $this->internetPrinter;
+        }
+
         if (null !== $this->localDrive) {
             $res['LocalDrive'] = $this->localDrive;
         }
@@ -993,6 +1009,10 @@ class describePolicyGroups extends Model
 
         if (null !== $this->resourceGroupCount) {
             $res['ResourceGroupCount'] = $this->resourceGroupCount;
+        }
+
+        if (null !== $this->safeMenu) {
+            $res['SafeMenu'] = $this->safeMenu;
         }
 
         if (null !== $this->scope) {
@@ -1296,6 +1316,10 @@ class describePolicyGroups extends Model
             $model->internetCommunicationProtocol = $map['InternetCommunicationProtocol'];
         }
 
+        if (isset($map['InternetPrinter'])) {
+            $model->internetPrinter = $map['InternetPrinter'];
+        }
+
         if (isset($map['LocalDrive'])) {
             $model->localDrive = $map['LocalDrive'];
         }
@@ -1474,6 +1498,10 @@ class describePolicyGroups extends Model
 
         if (isset($map['ResourceGroupCount'])) {
             $model->resourceGroupCount = $map['ResourceGroupCount'];
+        }
+
+        if (isset($map['SafeMenu'])) {
+            $model->safeMenu = $map['SafeMenu'];
         }
 
         if (isset($map['Scope'])) {